refactor(core): извлечь ядро audio + контекст и card-ownership хелпер
